32-bit mingw32 compile fix

+++ make-win-devtest.py
@@ -44,19 +44,22 @@ libDir = os.path.join("code", "libs", "win64")
 buildDir = os.path.join("build", "release-mingw32-x86_64")
 wolfcamDir = os.path.join(packageDir, "wolfcam-ql")
 
-print("copying 64-bit Windows binaries...")
-try:
-    shutil.copy2(os.path.join(libDir, "SDL264.dll"), packageDir)
-    shutil.copy2(os.path.join(libDir, "backtrace64.dll"), packageDir)
-    shutil.copy2(os.path.join(buildDir, "ioquake3.x86_64.exe"), packageDir)
-    shutil.move(os.path.join(packageDir, "ioquake3.x86_64.exe"), os.path.join(packageDir, "wolfcamql.exe"))
-    shutil.copy2(os.path.join(buildDir, "renderer_opengl1_x86_64.dll"), packageDir)
-    shutil.copy2(os.path.join(buildDir, "renderer_opengl2_x86_64.dll"), packageDir)
-    shutil.copy2(os.path.join(buildDir, "baseq3", "cgamex86_64.dll"), wolfcamDir)
-    shutil.copy2(os.path.join(buildDir, "baseq3", "qagamex86_64.dll"), wolfcamDir)
-    shutil.copy2(os.path.join(buildDir, "baseq3", "uix86_64.dll"), wolfcamDir)
-except IOError as err:
-    print(err)
+if os.path.exists(os.path.join(buildDir, "ioquake3.x86_64.exe")):
+    print("copying 64-bit Windows binaries...")
+    try:
+        shutil.copy2(os.path.join(libDir, "SDL264.dll"), packageDir)
+        shutil.copy2(os.path.join(libDir, "backtrace64.dll"), packageDir)
+        shutil.copy2(os.path.join(buildDir, "ioquake3.x86_64.exe"), packageDir)
+        shutil.move(os.path.join(packageDir, "ioquake3.x86_64.exe"), os.path.join(packageDir, "wolfcamql.exe"))
+        shutil.copy2(os.path.join(buildDir, "renderer_opengl1_x86_64.dll"), packageDir)
+        shutil.copy2(os.path.join(buildDir, "renderer_opengl2_x86_64.dll"), packageDir)
+        shutil.copy2(os.path.join(buildDir, "baseq3", "cgamex86_64.dll"), wolfcamDir)
+        shutil.copy2(os.path.join(buildDir, "baseq3", "qagamex86_64.dll"), wolfcamDir)
+        shutil.copy2(os.path.join(buildDir, "baseq3", "uix86_64.dll"), wolfcamDir)
+    except IOError as err:
+        print(err)
+else:
+    print("skipping 64-bit Windows binaries")
 
 # 32-bit
 
@@ -64,19 +67,22 @@ libDir = os.path.join("code", "libs", "win32")
 buildDir = os.path.join("build", "release-mingw32-x86")
 wolfcamDir = os.path.join(packageDir, "wolfcam-ql")
 
-print("copying 32-bit Windows binaries...")
-try:
-    shutil.copy2(os.path.join(libDir, "SDL2.dll"), packageDir)
-    shutil.copy2(os.path.join(libDir, "backtrace.dll"), packageDir)
-    shutil.copy2(os.path.join(buildDir, "ioquake3.x86.exe"), packageDir)
-    shutil.move(os.path.join(packageDir, "ioquake3.x86.exe"), os.path.join(packageDir, "wolfcamql32.exe"))
-    shutil.copy2(os.path.join(buildDir, "renderer_opengl1_x86.dll"), packageDir)
-    shutil.copy2(os.path.join(buildDir, "renderer_opengl2_x86.dll"), packageDir)
-    shutil.copy2(os.path.join(buildDir, "baseq3", "cgamex86.dll"), wolfcamDir)
-    shutil.copy2(os.path.join(buildDir, "baseq3", "qagamex86.dll"), wolfcamDir)
-    shutil.copy2(os.path.join(buildDir, "baseq3", "uix86.dll"), wolfcamDir)
-except IOError as err:
-    print(err)
+if os.path.exists(os.path.join(buildDir, "ioquake3.x86.exe")):
+    print("copying 32-bit Windows binaries...")
+    try:
+        shutil.copy2(os.path.join(libDir, "SDL2.dll"), packageDir)
+        shutil.copy2(os.path.join(libDir, "backtrace.dll"), packageDir)
+        shutil.copy2(os.path.join(buildDir, "ioquake3.x86.exe"), packageDir)
+        shutil.move(os.path.join(packageDir, "ioquake3.x86.exe"), os.path.join(packageDir, "wolfcamql32.exe"))
+        shutil.copy2(os.path.join(buildDir, "renderer_opengl1_x86.dll"), packageDir)
+        shutil.copy2(os.path.join(buildDir, "renderer_opengl2_x86.dll"), packageDir)
+        shutil.copy2(os.path.join(buildDir, "baseq3", "cgamex86.dll"), wolfcamDir)
+        shutil.copy2(os.path.join(buildDir, "baseq3", "qagamex86.dll"), wolfcamDir)
+        shutil.copy2(os.path.join(buildDir, "baseq3", "uix86.dll"), wolfcamDir)
+    except IOError as err:
+        print(err)
+else:
+    print("skipping 32-bit Windows binaries")
 
 print("copying misc files...")
 shutil.copy2(os.path.join("ui", "wcmenudef.h"), os.path.join(packageDir, "wolfcam-ql", "ui"))
